Extract model account_move_line
