+++ account_constraints/model/__init__.py
@@ -2,3 +2,4 @@
 
 from . import account_invoice
 from . import account_bank_statement
+from . import account_move_line
